+++ targon/utils/__init__.py
@@ -1,5 +1,4 @@
 from . import config
-from . import uids
 from . import updater
 from . import misc
 from . import prompt
