+++ targon/verifier/__init__.py
@@ -3,3 +3,4 @@ from . import challenge
 from . import database
 from . import inference
 from . import state
+from . import uids
